+++ js/functions.js
@@ -5,17 +5,18 @@ function checkLenght (string, maxLength) {
     return false;
   }
 }
-console.log(checkLenght('проверяемая строка', 10));
+checkLenght('проверяемая строка', 10);
 
 function isPalingrome (string) {
   string = string.replaceAll(' ', '');
   string = string.toLowerCase();
-  reverseString = '';
+  let reverseString = '';
   for (let i = string.length - 1; i >= 0; i--) {
-      reverseString+=string[i];
+    reverseString += string[i];
   }
   if (reverseString === string) {
-      return true;
+    return true;
   }
   return false;
 }
+isPalingrome('fkkf');
